Logica km autista - inserimento multiplo

+++ src/session/infrastructure/persistence/session.repository.ts
@@ -5,23 +5,16 @@ import { Session } from '../../domain/session';
 export abstract class SessionRepository {
   abstract findById(id: Session['id']): Promise<NullableType<Session>>;
 
-  abstract create(
-    data: Omit<Session, 'id' | 'createdAt' | 'updatedAt' | 'deletedAt' | 'user.operatori'>,
-  ): Promise<Session>;
+  abstract create(data: Omit<Session, 'id' | 'createdAt' | 'updatedAt' | 'deletedAt' | 'user.operatori'>): Promise<Session>;
 
   abstract update(
     id: Session['id'],
-    payload: Partial<
-      Omit<Session, 'id' | 'createdAt' | 'updatedAt' | 'deletedAt'>
-    >,
+    payload: Partial<Omit<Session, 'id' | 'createdAt' | 'updatedAt' | 'deletedAt'>>,
   ): Promise<Session | null>;
 
   abstract deleteById(id: Session['id']): Promise<void>;
 
   abstract deleteByUserId(conditions: { userId: User['id'] }): Promise<void>;
 
-  abstract deleteByUserIdWithExclude(conditions: {
-    userId: User['id'];
-    excludeSessionId: Session['id'];
-  }): Promise<void>;
+  abstract deleteByUserIdWithExclude(conditions: { userId: User['id']; excludeSessionId: Session['id'] }): Promise<void>;
 }
